Split into core submodule
